requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# LangChain
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-community==0.3.7
langchain-core==0.3.15

# LangSmith (Tracing & Monitoring)
langsmith==0.1.137

# Supabase
supabase==2.10.0

# Embeddings & Vector
numpy==1.26.4

# HTTP
httpx==0.27.2
